Refactoring sensor code to pipe it through the telemetry system.
