+++ .gitignore
@@ -1,3 +1,5 @@
 build/
 .yotta*
 yotta_*
+.vscode
+.kubos*
