+++ source/disk.h
@@ -1,0 +1,8 @@
+#ifndef DISK_H
+#define DISK_H
+
+#include <kubos-core/modules/fs/fs.h>
+#include <kubos-core/modules/fatfs/ff.h>
+#include <kubos-core/modules/fatfs/diskio.h>
+
+#endif
