+++ source/sensor.h
@@ -1,0 +1,8 @@
+#ifndef SENSOR_H
+#define SENSOR_H
+
+#include <csp/arch/csp_thread.h>
+
+CSP_DEFINE_TASK(calibrate_thread);
+
+#endif
